Add packages to Java-Kafka directories (#115)
* Add packages to Java-Kafka directories


* Add packages to Java-Kafka directories


* Add packages to Java-Kafka directories


* Remove empty line below copyright


---------
