+++ java/common/src/main/java/io/strimzi/common/TracingSystem.java
@@ -2,6 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
+package io.strimzi.common;
 
 public enum TracingSystem {
     JAEGER,
