+++ java/kafka/consumer/pom.xml
@@ -83,7 +83,7 @@
                             </filters>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>KafkaConsumerExample</mainClass>
+                                    <mainClass>io.strimzi.kafka.consumer.KafkaConsumerExample</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
