+++ java/kafka/consumer/src/main/java/io/strimzi/kafka/consumer/KafkaConsumerExample.java
@@ -2,7 +2,10 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
+package io.strimzi.kafka.consumer;
 
+import io.strimzi.common.TracingInitializer;
+import io.strimzi.common.TracingSystem;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
