+++ java/kafka/producer/pom.xml
@@ -83,7 +83,7 @@
                             </filters>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>KafkaProducerExample</mainClass>
+                                    <mainClass>io.strimzi.kafka.producer.KafkaProducerExample</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
