+++ java/kafka/producer/src/main/java/io/strimzi/kafka/producer/KafkaProducerConfig.java
@@ -2,8 +2,10 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
+package io.strimzi.kafka.producer;
 
 import io.strimzi.common.ConfigUtil;
+import io.strimzi.common.TracingSystem;
 
 import java.util.Map;
 import java.util.Properties;
