+++ java/kafka/producer/src/main/java/io/strimzi/kafka/producer/KafkaProducerExample.java
@@ -2,7 +2,10 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
+package io.strimzi.kafka.producer;
 
+import io.strimzi.common.TracingInitializer;
+import io.strimzi.common.TracingSystem;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
