+++ java/kafka/streams/src/main/java/io/strimzi/kafka/streams/TracingKafkaClientSupplier.java
@@ -2,6 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
+package io.strimzi.kafka.streams;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.instrumentation.kafkaclients.KafkaTelemetry;
